precision_landing/requirements.txt
opencv-contrib-python>=4.7.0
numpy>=1.21.0
pymavlink>=2.4.38
pyyaml>=6.0
pupil-apriltags>=1.0.4; platform_machine != 'armv7l' and platform_machine != 'aarch64' # Optional logic per architecture
